requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
distro==1.7.0
Django==4.0.3
django-cors-headers==3.11.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fastjsonschema==2.15.3
fonttools==4.31.2
fuzzywuzzy==0.18.0
gunicorn==20.1.0
idna==3.3
ipykernel==6.12.1
ipython==8.2.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.1
jsonschema==4.4.0
jupyter-client==7.2.1
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.13
nbconvert==6.4.5
nbformat==5.3.0
nest-asyncio==1.5.5
notebook==6.4.10
numpy==1.22.3
oauthlib==3.2.0
openpyxl==3.0.9
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.0
prometheus-client==0.14.0
prompt-toolkit==3.0.29
psutil==5.9.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
pymemcache==3.5.2
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-Levenshtein==0.12.2
python-magic==0.4.25
python-magic-bin @ file:///C:/Users/JHET/Documents/TRASH_CODES/brngy_census_analyzer/python_magic_bin-0.4.14-py2.py3-none-win_amd64.whl
python3-openid==3.2.0
pytz==2022.1
pywin32==303
pywinpty==2.0.5
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.1
Send2Trash==1.8.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
soupsieve==2.3.2
sqlparse==0.4.2
stack-data==0.2.0
tabula-py==2.3.0
terminado==0.13.3
testpath==0.6.0
tornado==6.1
traitlets==5.1.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
